Etc: pageViewControllerDelegate 공통 사항 CommonProductPageViewControllerRenderable로 분리
