+++ Pyonsnal-Color/Pyonsnal-Color/Common/Protocol/CommonProductPageViewControllerRenderable.swift
@@ -1,0 +1,17 @@
+//
+//  CommonProductPageViewControllerRenderable.swift
+//  Pyonsnal-Color
+//
+//  Created by 조소정 on 2023/09/09.
+//
+
+import Foundation
+
+protocol CommonProductPageViewControllerRenderable: AnyObject {
+    func updateSelectedStoreCell(index: Int)
+    func didChangeStore(to store: ConvenienceStore)
+    func didSelect(with brandProduct: ProductConvertable)
+    func deleteKeywordFilter(_ filter: FilterItemEntity)
+    func didTapRefreshFilterButton()
+    func didFinishUpdateSnapshot()
+}
